Playground code clean up (#2895)
