+++ apps/docsite/src/app/playground/PlaygroundClient.tsx
@@ -13,7 +13,7 @@
  *   - Property: component selector + instance picker + knobs that edit the code.
  * Right panel: toolbar (dark mode · target element · viewport
  *   segmented control · share · expand) over a responsive
- *   /playground-preview iframe.
+ *   /playground/preview iframe.
  *
  * The preview iframe is driven via postMessage (preview-code / preview-theme);
  * code lives in React state and is the single source of truth shared by Monaco,
@@ -25,13 +25,14 @@
 import {useCallback, useEffect, useMemo, useRef, useState} from 'react';
 import {useSearchParams} from 'next/navigation';
 import dynamic from 'next/dynamic';
-import {loader} from '@monaco-editor/react';
 import * as stylex from '@stylexjs/stylex';
 import {XDSAppShell} from '@xds/core/AppShell';
 import {compressCode, decompressCode} from '../../lib/compress';
 import {XDSButton} from '@xds/core/Button';
 import {XDSLink} from '@xds/core/Link';
 import {XDSHStack, XDSVStack} from '@xds/core/Layout';
+import {XDSCenter} from '@xds/core/Center';
+import {XDSSpinner} from '@xds/core/Spinner';
 import {XDSPopover} from '@xds/core/Popover';
 import {XDSTextInput} from '@xds/core/TextInput';
 import {
@@ -68,74 +69,41 @@ import {
   RotateCw,
   Crosshair,
 } from 'lucide-react';
-import githubLight from './themes/github-light.json';
-import githubDark from './themes/github-dark.json';
+import githubLight from './codeEditorThemes/github-light.json';
+import githubDark from './codeEditorThemes/github-dark.json';
 import {useThemeMode} from '../providers';
 import {
   DEFAULT_PLAYGROUND_THEME,
   PLAYGROUND_THEME_OPTIONS,
   themeByValue,
-} from './playgroundThemes';
+} from './previewThemes';
 import {templates} from '../../generated/templateRegistry';
 import {PreviewStage, type Viewport} from './PreviewStage';
 import {ConfirmDialog} from './ConfirmDialog';
 import {BRAND_ICON} from '../../components/XDSWordmark';
-import {annotateInstanceIds} from './babelParser';
+import {annotateInstanceIds} from './propertyEditor/componentInstances';
 import {trackCopy} from '../../lib/analytics';
-import {PlaygroundThemeEditor} from '../../components/themePlayground/PlaygroundThemeEditor';
-import {generateThemeCode} from '../../components/themePlayground/helpers';
+import {ThemeEditor} from './themeEditor/ThemeEditor';
+import {generateThemeCode} from './themeEditor/helpers';
 import {DEFAULT_CODE} from './defaultCode';
 import {stripCodeExampleCopyrightHeader} from '../../lib/codeExamples';
+import {configureMonaco, type MonacoInstance} from './monacoSetup';
 
 import type * as MonacoTypes from 'monaco-editor';
 import type {XDSDefinedTheme} from '@xds/core/theme';
 import {xdsTokenDefaults} from '@xds/core/theme';
 
-// Source of the theme-showcase template — a set of real product surfaces
-// (store, checkout, chat, inventory) used to preview a theme. It's hidden
-// from the generated templates registry (isHiddenFromOverview), so it's added
+// Hidden from the generated registry (isHiddenFromOverview), so it's added
 // explicitly as the first entry in the Templates dropdown.
 const THEME_SHOWCASE_SOURCE =
   templates.find(t => t.slug === 'theme-showcase')?.source ?? '';
 
-// Self-host Monaco from public/monaco/vs — corpnet blocks the default
-// jsdelivr CDN. Configure the singleton loader before the editor initializes.
-if (typeof window !== 'undefined') {
-  loader.config({paths: {vs: '/monaco/vs'}});
-}
-
-/** Monaco instance type — the full runtime object passed to onMount */
-type MonacoInstance = typeof MonacoTypes & {
-  languages: typeof MonacoTypes.languages & {
-    typescript: {
-      typescriptDefaults: {
-        setCompilerOptions: (options: Record<string, unknown>) => void;
-        setDiagnosticsOptions: (options: Record<string, unknown>) => void;
-        addExtraLib: (content: string, filePath: string) => void;
-      };
-      ScriptTarget: Record<string, number>;
-      ModuleKind: Record<string, number>;
-      JsxEmit: Record<string, number>;
-      ModuleResolutionKind: Record<string, number>;
-    };
-  };
-  editor: typeof MonacoTypes.editor & {
-    defineTheme: (name: string, data: Record<string, unknown>) => void;
-  };
-};
-
 const MonacoEditor = dynamic(() => import('@monaco-editor/react'), {
   ssr: false,
   loading: () => (
-    <div
-      style={{
-        flex: 1,
-        display: 'flex',
-        alignItems: 'center',
-        justifyContent: 'center',
-      }}>
-      <XDSText color="secondary">Loading editor…</XDSText>
-    </div>
+    <XDSCenter height="100%">
+      <XDSSpinner label="Loading editor…" />
+    </XDSCenter>
   ),
 });
 
@@ -162,128 +130,6 @@ function getInitialCode(): string {
 function updateURL(code: string) {
   const compressed = compressCode(code);
   window.history.replaceState(null, '', `#code=${compressed}`);
-}
-
-/**
- * Configure Monaco's TypeScript service with real XDS type definitions.
- * Loads .d.ts files from a pre-built JSON bundle (generated at build time).
- */
-function configureMonaco(monaco: MonacoInstance) {
-  const ts = monaco.languages.typescript.typescriptDefaults;
-
-  ts.setCompilerOptions({
-    target: monaco.languages.typescript.ScriptTarget.ESNext,
-    module: monaco.languages.typescript.ModuleKind.ESNext,
-    jsx: monaco.languages.typescript.JsxEmit.ReactJSX,
-    esModuleInterop: true,
-    allowSyntheticDefaultImports: true,
-    moduleResolution: monaco.languages.typescript.ModuleResolutionKind.NodeJs,
-    allowJs: true,
-    strict: false,
-  });
-
-  ts.setDiagnosticsOptions({
-    noSemanticValidation: true,
-    noSyntaxValidation: false,
-  });
-
-  // Global declarations for React hooks (available without import in playground)
-  ts.addExtraLib(
-    `declare function useState<T>(init: T | (() => T)): [T, (v: T | ((prev: T) => T)) => void];
-    declare function useEffect(fn: () => void | (() => void), deps?: readonly unknown[]): void;
-    declare function useCallback<T extends Function>(fn: T, deps: readonly unknown[]): T;
-    declare function useMemo<T>(fn: () => T, deps: readonly unknown[]): T;
-    declare function useRef<T>(init: T): { current: T };
-    declare function useReducer<S, A>(reducer: (state: S, action: A) => S, init: S): [S, (action: A) => void];
-    declare function useContext<T>(ctx: unknown): T;`,
-    'file:///globals.d.ts',
-  );
-
-  // Lucide icons wildcard stub — gives Monaco a sense of the
-  // module's shape so import { Icon } from 'lucide-react' doesn't
-  // light up with red squigglies in the playground editor.
-  ts.addExtraLib(
-    `declare module 'lucide-react' { const icons: Record<string, React.ComponentType<{size?: number | string; color?: string; strokeWidth?: number | string; className?: string}>>; export = icons; }`,
-    'file:///node_modules/lucide-react/index.d.ts',
-  );
-
-  // Load real type definitions from the pre-built JSON bundle
-  fetch('/playground-types.json')
-    .then(r => r.json())
-    .then((packages: Record<string, Record<string, string>>) => {
-      const reactFiles = packages['react'] ?? {};
-      for (const [fileName, content] of Object.entries(reactFiles)) {
-        ts.addExtraLib(
-          content,
-          `file:///node_modules/@types/react/${fileName}`,
-        );
-        // Also register react/jsx-runtime as a resolvable module path
-        if (fileName === 'jsx-runtime.d.ts') {
-          ts.addExtraLib(
-            content,
-            'file:///node_modules/react/jsx-runtime.d.ts',
-          );
-        }
-      }
-
-      const stylexFiles = packages['@stylexjs/stylex'] ?? {};
-      for (const [fileName, content] of Object.entries(stylexFiles)) {
-        ts.addExtraLib(
-          content,
-          `file:///node_modules/@stylexjs/stylex/${fileName}`,
-        );
-      }
-
-      // Heroicons ambient declarations, one per size/style variant. Template
-      // and example code imports icons by name from
-      // '@heroicons/react/{16,20,24}/{outline,solid}', so each declaration
-      // exposes the variant's icons as named exports. Without these, every
-      // heroicons import lights up with a "Cannot find module" red squiggle
-      // once semantic validation turns on below.
-      const heroiconFiles = packages['@heroicons/react'] ?? {};
-      for (const [fileName, content] of Object.entries(heroiconFiles)) {
-        const variant = fileName.replace(/\.d\.ts$/, '');
-        ts.addExtraLib(
-          content,
-          `file:///node_modules/@heroicons/react/${variant}/index.d.ts`,
-        );
-      }
-
-      const xdsFiles = packages['@xds/core'] ?? {};
-      const submoduleReexports: string[] = [];
-
-      for (const [relPath, content] of Object.entries(xdsFiles)) {
-        ts.addExtraLib(
-          content,
-          `file:///node_modules/@xds/core/dist/${relPath}`,
-        );
-
-        if (relPath.endsWith('/index.d.ts')) {
-          const moduleName = relPath.replace('/index.d.ts', '');
-          ts.addExtraLib(
-            content,
-            `file:///node_modules/@xds/core/${moduleName}/index.d.ts`,
-          );
-          submoduleReexports.push(moduleName);
-        }
-      }
-
-      const barrelContent = submoduleReexports
-        .map(m => `export * from '@xds/core/${m}';`)
-        .join('\n');
-      ts.addExtraLib(
-        `declare module '@xds/core' {\n${barrelContent}\n}`,
-        'file:///node_modules/@xds/core/index.d.ts',
-      );
-
-      ts.setDiagnosticsOptions({
-        noSemanticValidation: false,
-        noSyntaxValidation: false,
-      });
-    })
-    .catch(() => {
-      // Types unavailable — keep semantic validation disabled
-    });
 }
 
 type LeftView = 'code' | 'theme';
@@ -336,21 +182,12 @@ const s = stylex.create({
   tabBody: {
     flex: 1,
     minHeight: 0,
-    display: 'flex',
-    flexDirection: 'column',
     overflow: 'hidden',
   },
-  codePane: {
+  pane: {
     flex: 1,
     minHeight: 0,
     minWidth: 0,
-  },
-  themePane: {
-    flex: 1,
-    minHeight: 0,
-    minWidth: 0,
-    display: 'flex',
-    flexDirection: 'column',
   },
   rightPanel: {
     flex: 1,
@@ -361,12 +198,12 @@ const s = stylex.create({
     backgroundColor: 'var(--color-background-muted)',
   },
   buildStatus: {
-    display: 'flex',
-    alignItems: 'center',
-    gap: 'var(--spacing-2)',
     transitionProperty: 'opacity',
     transitionDuration: '0.5s',
     transitionTimingFunction: 'ease',
+  },
+  buildStatusFaded: {
+    opacity: 0,
   },
   sideNavHeading: {
     paddingInline: {
@@ -395,22 +232,17 @@ export function PlaygroundClient({defaultIsMobile}: PlaygroundClientProps) {
   const [code, setCode] = useState(getInitialCode);
   const [mode, setMode] = useState<'light' | 'dark'>('light');
   // A ?theme=<value> query param (e.g. from the themes gallery's "Open in
-  // Playground") seeds the Theme view with that theme and applies it to the
-  // preview. useSearchParams reads it reliably across App Router client-side
-  // navigation (a window.location read at mount can be stale on soft nav).
-  // Validated against the registered themes; null when absent or unknown so
-  // the playground keeps its default, unseeded behavior.
+  // Playground") seeds the Theme view and preview. useSearchParams reads it
+  // reliably across soft navigation; validated against registered themes.
   const searchParams = useSearchParams();
   const rawThemeParam = searchParams.get('theme');
   const themeParam =
     rawThemeParam && rawThemeParam in themeByValue ? rawThemeParam : null;
   const theme = themeParam ?? DEFAULT_PLAYGROUND_THEME;
-  // The theme whose values seed the Theme editor: the ?theme= theme on first
-  // load, then whichever theme the user picks from the "Themes" dropdown.
-  // Changing it remounts the editor (via key) so it re-populates from that theme.
+  // The theme that seeds the Theme editor: the ?theme= theme on first load, then
+  // whichever theme the user picks from "Themes". Changing it remounts the
+  // editor (via key) so it re-populates.
   const [selectedTheme, setSelectedTheme] = useState(theme);
-  // The theme object the editor is seeded from. Always defined so the always-
-  // mounted Theme editor reflects the active theme and stays the source of truth.
   const editorInitialTheme =
     themeByValue[selectedTheme] ?? themeByValue[DEFAULT_PLAYGROUND_THEME];
   const [activeView, setActiveView] = useState<LeftView>('code');
@@ -423,26 +255,18 @@ export function PlaygroundClient({defaultIsMobile}: PlaygroundClientProps) {
   const [copied, setCopied] = useState(false);
   const [shareUrl, setShareUrl] = useState('');
   const [themeName, setThemeName] = useState('');
-  // Snapshot of the theme to export, captured when the Export popover opens.
-  // The live theme lives in customThemeRef (a ref, updated as the user edits
-  // tokens without re-rendering the playground); we copy it into state on
-  // open so the download link below can be built declaratively from reactive
-  // state. The popover is in the toolbar, so the theme isn't being edited
-  // while it's open — the snapshot stays current.
+  // Snapshot of the theme to export, captured when the Export popover opens, so
+  // the download link can be built declaratively from reactive state (the live
+  // theme lives in customThemeRef and changes without re-rendering).
   const [exportTheme, setExportTheme] = useState<XDSDefinedTheme | null>(null);
   const [previewReady, setPreviewReady] = useState(false);
   const [isTargeting, setIsTargeting] = useState(false);
-  // The theme value awaiting confirmation from the "Example themes" dropdown.
-  // Applying an example theme re-seeds the Theme editor and discards the
-  // current theme, so we hold the pending choice until the user confirms.
-  // The dialog is open whenever this is non-null.
+  // Pending confirmations: applying an example theme / loading a template both
+  // discard the user's current work, so the choice is held until they confirm.
+  // The matching dialog is open whenever the value is non-null.
   const [pendingExampleTheme, setPendingExampleTheme] = useState<string | null>(
     null,
   );
-  // The template source awaiting confirmation from the "Templates" dropdown.
-  // Loading a template overwrites the code editor, so we hold the pending
-  // choice until the user confirms. The dialog is open whenever this is
-  // non-null.
   const [pendingTemplateSource, setPendingTemplateSource] = useState<
     string | null
   >(null);
@@ -454,12 +278,10 @@ export function PlaygroundClient({defaultIsMobile}: PlaygroundClientProps) {
   const editorRef = useRef<MonacoTypes.editor.IStandaloneCodeEditor | null>(
     null,
   );
-  // Points at the declaratively-rendered theme-export download link (below).
-  // The Download button triggers it so we get native download behavior from a
-  // normally-rendered <a> rather than fabricating one in a click handler.
+  // Points at the hidden theme-export download link (below); the Download button
+  // clicks it for native download behavior from a real <a>.
   const downloadLinkRef = useRef<HTMLAnchorElement>(null);
-  // Mirror activeView in a ref so onMount can read the current view without
-  // re-creating the (stable) mount callback.
+  // Mirror activeView in a ref so the stable onMount callback can read it.
   const activeViewRef = useRef(activeView);
   activeViewRef.current = activeView;
   // Latest theme authored in the Theme view, retained so a mode toggle can
@@ -473,11 +295,8 @@ export function PlaygroundClient({defaultIsMobile}: PlaygroundClientProps) {
     autoSaveId: 'xds-playground-left-width',
   });
 
-  // Seed the initial preview mode from the docsite's mode so the preview opens
-  // consistent with the rest of the site. (The toolbar toggle can still switch
-  // the preview independently afterward.)
-  // Seed once on mount with the current docsite mode; afterward the preview
-  // toggle owns this state, so siteMode is intentionally read only initially.
+  // Seed the preview mode once from the docsite's mode; afterward the toolbar
+  // toggle owns it, so siteMode is intentionally read only on mount.
   const initialSiteModeRef = useRef(siteMode);
   useEffect(() => {
     setMode(initialSiteModeRef.current);
@@ -495,14 +314,10 @@ export function PlaygroundClient({defaultIsMobile}: PlaygroundClientProps) {
     return () => window.removeEventListener('hashchange', onHashChange);
   }, []);
 
-  // Re-read the hash once on mount. When arriving via a soft (client-side)
-  // navigation — e.g. the Themes page's "Open in Playground" link routed
-  // through Next's <Link> — the App Router commits the new URL slightly
-  // after this component first renders, so the synchronous getInitialCode()
-  // seed for `code` above can miss the fragment and fall back to
-  // DEFAULT_CODE. A hashchange event isn't emitted for that navigation
-  // (the page itself changes, not just the fragment), so we read the
-  // committed hash here, after mount, and adopt the seeded code if present.
+  // Re-read the hash once on mount. On soft navigation (e.g. the Themes page's
+  // "Open in Playground" <Link>) the App Router commits the URL after first
+  // render, so the synchronous seed above can miss the fragment, and no
+  // hashchange fires. Adopt the committed hash's code here if present.
   useEffect(() => {
     const seeded = getInitialCode();
     if (seeded !== DEFAULT_CODE) {
@@ -521,16 +336,22 @@ export function PlaygroundClient({defaultIsMobile}: PlaygroundClientProps) {
     }
   }, [themeParam]);
 
-  const send = useCallback((c: string, source?: string) => {
-    const win = iframeRef.current?.contentWindow;
-    if (!win) {
-      return;
-    }
-    win.postMessage(
-      c ? {type: 'preview-code', code: c, source} : {type: 'preview-clear'},
+  // Single channel to the preview iframe; no-ops until the iframe exists.
+  const postToPreview = useCallback((message: unknown) => {
+    iframeRef.current?.contentWindow?.postMessage(
+      message,
       window.location.origin,
     );
   }, []);
+
+  const send = useCallback(
+    (c: string, source?: string) => {
+      postToPreview(
+        c ? {type: 'preview-code', code: c, source} : {type: 'preview-clear'},
+      );
+    },
+    [postToPreview],
+  );
 
   // Send the preview an instance-annotated copy of the code (markers let the
   // preview map a selected component to its DOM node for the focus-ring flash),
@@ -546,29 +367,26 @@ export function PlaygroundClient({defaultIsMobile}: PlaygroundClientProps) {
   const postCustomTheme = useCallback(
     (customTheme: XDSDefinedTheme) => {
       customThemeRef.current = customTheme;
-      iframeRef.current?.contentWindow?.postMessage(
-        {
-          type: 'preview-theme',
-          customTokens: customTheme.tokens,
-          customComponents: customTheme.components,
-          mode,
-        },
-        window.location.origin,
-      );
+      postToPreview({
+        type: 'preview-theme',
+        customTokens: customTheme.tokens,
+        customComponents: customTheme.components,
+        mode,
+      });
     },
-    [mode],
+    [mode, postToPreview],
   );
 
-  const toggleTargeting = useCallback((pressed?: boolean) => {
-    setIsTargeting(prev => {
-      const next = pressed ?? !prev;
-      iframeRef.current?.contentWindow?.postMessage(
-        {type: next ? 'targeting-enable' : 'targeting-disable'},
-        window.location.origin,
-      );
-      return next;
-    });
-  }, []);
+  const toggleTargeting = useCallback(
+    (pressed?: boolean) => {
+      setIsTargeting(prev => {
+        const next = pressed ?? !prev;
+        postToPreview({type: next ? 'targeting-enable' : 'targeting-disable'});
+        return next;
+      });
+    },
+    [postToPreview],
+  );
 
   useEffect(() => {
     const handler = (e: MessageEvent) => {
@@ -600,10 +418,7 @@ export function PlaygroundClient({defaultIsMobile}: PlaygroundClientProps) {
         // mode. Properties are edited via the badge's popover now, so the left
         // panel stays on whatever view is currently open.
         setIsTargeting(false);
-        iframeRef.current?.contentWindow?.postMessage(
-          {type: 'targeting-disable'},
-          window.location.origin,
-        );
+        postToPreview({type: 'targeting-disable'});
       }
       if (e.data?.type === 'targeting-exit') {
         setIsTargeting(false);
@@ -611,7 +426,7 @@ export function PlaygroundClient({defaultIsMobile}: PlaygroundClientProps) {
     };
     window.addEventListener('message', handler);
     return () => window.removeEventListener('message', handler);
-  }, [postCode]);
+  }, [postCode, postToPreview]);
 
   useEffect(() => {
     const interval = setInterval(() => {
@@ -619,13 +434,10 @@ export function PlaygroundClient({defaultIsMobile}: PlaygroundClientProps) {
         clearInterval(interval);
         return;
       }
-      iframeRef.current?.contentWindow?.postMessage(
-        {type: 'preview-ping'},
-        window.location.origin,
-      );
+      postToPreview({type: 'preview-ping'});
     }, 300);
     return () => clearInterval(interval);
-  }, []);
+  }, [postToPreview]);
 
   // Debounced push of code → preview + URL hash
   useEffect(() => {
@@ -650,35 +462,17 @@ export function PlaygroundClient({defaultIsMobile}: PlaygroundClientProps) {
     };
   }, [code, postCode]);
 
-  // Theme + mode → preview. Also re-sent once the preview signals ready so a
-  // non-default initial theme (e.g. neutral) applies even if this effect first
-  // ran before the iframe was listening. Once the Theme editor has produced a
-  // theme (seeded via ?theme= or edited by the user), it becomes the source of
-  // truth and is applied across ALL views (Code / Properties / Theme) so the
-  // preview stays consistent when switching tabs and when toggling light/dark.
+  // Theme + mode → preview. Re-sent on previewReady so a non-default initial
+  // theme still applies if this first ran before the iframe was listening. A
+  // theme authored in the Theme editor takes precedence and stays the source of
+  // truth across all views.
   useEffect(() => {
-    const win = iframeRef.current?.contentWindow;
-    if (!win) {
-      return;
-    }
     if (customThemeRef.current) {
-      const custom = customThemeRef.current;
-      win.postMessage(
-        {
-          type: 'preview-theme',
-          customTokens: custom.tokens,
-          customComponents: custom.components,
-          mode,
-        },
-        window.location.origin,
-      );
+      postCustomTheme(customThemeRef.current);
     } else {
-      win.postMessage(
-        {type: 'preview-theme', theme, mode},
-        window.location.origin,
-      );
+      postToPreview({type: 'preview-theme', theme, mode});
     }
-  }, [mode, previewReady, activeView, theme]);
+  }, [mode, previewReady, activeView, theme, postCustomTheme, postToPreview]);
 
   // While the resize handle is dragged, the cursor can pass over the preview
   // iframe — a separate document that swallows pointer events and stalls the
@@ -739,13 +533,9 @@ export function PlaygroundClient({defaultIsMobile}: PlaygroundClientProps) {
     });
   }, []);
 
-  // Derive the export href from reactive state so the download is a real,
-  // normally-rendered <a> (see downloadLinkRef below) rather than a DOM
-  // element fabricated and clicked inside a callback. The active theme is the
-  // one authored in the Theme editor (snapshotted into state when the popover
-  // opens), falling back to the seeded theme. Returns the suggested filename
-  // and a data: URL holding the generated `defineTheme` source; null until a
-  // theme name is entered.
+  // Derive the export href (filename + data: URL of the generated defineTheme
+  // source) from reactive state, so the download is a real <a> (downloadLinkRef)
+  // rather than a fabricated element. Null until a theme name is entered.
   const themeExport = useMemo(() => {
     const name = themeName.trim();
     if (!name) {
@@ -796,11 +586,8 @@ export function PlaygroundClient({defaultIsMobile}: PlaygroundClientProps) {
     return () => cancelAnimationFrame(id);
   }, [activeView]);
 
-  // Dropdown items for "Themes" — derived from the registered playground
-  // themes (PLAYGROUND_THEME_OPTIONS) so any installed @xds/theme-* package
-  // shows up automatically. Selecting one prompts for confirmation (it discards
-  // the current theme) before re-seeding the Theme editor (and thus the preview)
-  // with that theme's values.
+  // "Themes" dropdown — every registered playground theme. Selecting one prompts
+  // for confirmation before re-seeding the Theme editor (and preview).
   const themeMenuItems = useMemo(
     () =>
       PLAYGROUND_THEME_OPTIONS.map(option => ({
@@ -810,11 +597,8 @@ export function PlaygroundClient({defaultIsMobile}: PlaygroundClientProps) {
     [],
   );
 
-  // Dropdown items for "Templates" — the published, ready templates from the
-  // generated registry (the same source the /templates gallery renders), so any
-  // new template shows up automatically. Selecting one prompts for confirmation
-  // (it overwrites the editor) before loading its source (setCode drives Monaco,
-  // the preview, and the URL hash).
+  // "Templates" dropdown — the published, ready templates from the generated
+  // registry. Selecting one prompts for confirmation before loading its source.
   const templateMenuItems = useMemo(
     () => [
       {
@@ -852,8 +636,7 @@ export function PlaygroundClient({defaultIsMobile}: PlaygroundClientProps) {
       wordWrap: 'on' as const,
       padding: {top: 12},
       accessibilitySupport: 'off' as const,
-      // Hide scrollbars (wheel/keyboard scrolling still works) + the right-side
-      // overview ruler, for a cleaner panel.
+      // Hide scrollbars and the overview ruler for a cleaner panel.
       scrollbar: {
         vertical: 'hidden' as const,
         horizontal: 'hidden' as const,
@@ -1009,10 +792,10 @@ export function PlaygroundClient({defaultIsMobile}: PlaygroundClientProps) {
               </XDSHStack>
             </XDSHStack>
           )}
-          <div {...stylex.props(s.tabBody)}>
-            {/* Code: Monaco stays mounted to preserve typedefs + editor state */}
-            <div
-              {...stylex.props(s.codePane, activeView !== 'code' && s.hidden)}>
+          {/* Both panes stay mounted (hidden when inactive) so Monaco's
+              typedefs and the theme editor's state survive tab switches. */}
+          <XDSVStack xstyle={s.tabBody}>
+            <XDSVStack xstyle={[s.pane, activeView !== 'code' && s.hidden]}>
               <MonacoEditor
                 defaultLanguage="typescript"
                 value={code}
@@ -1023,24 +806,16 @@ export function PlaygroundClient({defaultIsMobile}: PlaygroundClientProps) {
                 onMount={handleMonacoMount}
                 options={editorOptions}
               />
-            </div>
-
-            {/* Theme: stays mounted (hidden when inactive) to preserve the
-                editor's token/component state across tab switches — it only
-                changes when the user edits it, not on navigation. */}
-            <div
-              {...stylex.props(
-                s.themePane,
-                activeView !== 'theme' && s.hidden,
-              )}>
-              <PlaygroundThemeEditor
+            </XDSVStack>
+            <XDSVStack xstyle={[s.pane, activeView !== 'theme' && s.hidden]}>
+              <ThemeEditor
                 key={selectedTheme}
                 mode={mode}
                 initialTheme={editorInitialTheme}
                 onThemeChange={postCustomTheme}
               />
-            </div>
-          </div>
+            </XDSVStack>
+          </XDSVStack>
         </XDSVStack>
 
         {!isMobile && (
@@ -1119,9 +894,13 @@ export function PlaygroundClient({defaultIsMobile}: PlaygroundClientProps) {
               endContent={
                 <XDSHStack gap={4} vAlign="center">
                   {buildStatus !== 'idle' && (
-                    <div
-                      {...stylex.props(s.buildStatus)}
-                      style={{opacity: statusFading ? 0 : 1}}>
+                    <XDSHStack
+                      gap={2}
+                      vAlign="center"
+                      xstyle={[
+                        s.buildStatus,
+                        statusFading && s.buildStatusFaded,
+                      ]}>
                       <XDSStatusDot
                         variant={BUILD_STATUS_META[buildStatus].variant}
                         label={BUILD_STATUS_META[buildStatus].label}
@@ -1141,7 +920,7 @@ export function PlaygroundClient({defaultIsMobile}: PlaygroundClientProps) {
                           onClick={handleRebuild}
                         />
                       )}
-                    </div>
+                    </XDSHStack>
                   )}
                   <XDSPopover
                     label="Share template"
