+++ apps/docsite/src/app/playground/codeEditorThemes/github-dark.json
