+++ apps/docsite/src/app/playground/defaultCode.ts
@@ -6,7 +6,7 @@
  * @position Playground — the initial example users see and edit.
  *
  * A minimal, self-contained XDSCard example (header, body text, footer button)
- * that runs in the preview sandbox (see (preview)/playground-preview/runner.ts).
+ * that runs in the preview sandbox (see playground/preview/runner.ts).
  * Kept as a string because the playground edits/compiles it as user code rather
  * than importing it.
  */
@@ -34,7 +34,7 @@ export default function Example() {
         content={
           <XDSLayoutContent>
             <XDSText type="body" color="secondary">
-              Try out components in the code editor, open a ready-made templates,
+              Try out components in the code editor, open a ready-made template,
               and build your own theme in the theme editor — all in one place.
             </XDSText>
           </XDSLayoutContent>
