+++ apps/docsite/src/app/playground/preview/previewReset.css
@@ -1,0 +1,13 @@
+/* Copyright (c) Meta Platforms, Inc. and affiliates. */
+
+/*
+ * Document reset for the preview iframe. Full-height html/body so user
+ * templates sized with height: 100% / 100dvh fill the frame. Imported by the
+ * preview route's layout.
+ */
+
+html,
+body {
+  height: 100%;
+  margin: 0;
+}
