+++ apps/docsite/src/app/playground/themeEditor/ThemeEditor.tsx
@@ -1,21 +1,22 @@
 // Copyright (c) Meta Platforms, Inc. and affiliates.
 
 /**
- * @file PlaygroundThemeEditor.tsx
+ * @file ThemeEditor.tsx
  * @input optional seed theme + the playground's light/dark mode
  * @output the theme editor's left-panel content (tab strip + scrollable body)
- * @position Playground — embeds the themePlayground editor as a left-panel view.
+ * @position Playground — the Theme tab's editor, rendered in the left panel.
  *
- * A self-contained version of the editor portion of ThemeEditorView: it owns the
- * token + scale state, renders the Theme/Components/Tokens tab strip and body, and
- * reports the composed theme upward via onThemeChange so the playground can push it
- * to the live preview iframe.
+ * A self-contained theme editor: it owns the token + scale state, renders the
+ * Theme/Components/Tokens tab strip and body, and reports the composed theme upward
+ * via onThemeChange so the playground can push it to the live preview iframe.
  */
 
 'use client';
 
-import * as React from 'react';
+import {useCallback, useEffect, useMemo, useState} from 'react';
+import * as stylex from '@stylexjs/stylex';
 import {XDSTabList, XDSTab} from '@xds/core/TabList';
+import {XDSVStack, XDSStackItem} from '@xds/core/Stack';
 import {
   defineTheme,
   expandTypeScale,
@@ -34,7 +35,7 @@ import {
   easeDefaults,
 } from '@xds/core/theme';
 import type {XDSDefinedTheme} from '@xds/core/theme';
-import {EditorSections} from './EditorSections';
+import {BaseStylesPanel} from './BaseStylesPanel';
 import {ComponentTokensPanel} from './ComponentTokensPanel';
 import type {CustomOverride} from './ComponentTokensPanel';
 import {RawTokensPanel} from './RawTokensPanel';
@@ -43,7 +44,24 @@ import {
   COMPONENT_VAR_NAMES,
   GOOGLE_FONTS_URL,
 } from './constants';
-import {buildComponentOverrides, mergeComponentStyleMaps} from './helpers';
+import {
+  buildComponentOverrides,
+  buildSpacingScale,
+  mergeComponentStyleMaps,
+} from './helpers';
+
+const s = stylex.create({
+  root: {
+    flex: 1,
+    minHeight: 0,
+    overflow: 'hidden',
+  },
+  body: {
+    flex: 1,
+    overflow: 'auto',
+    padding: 'var(--spacing-4)',
+  },
+});
 
 const ALL_DEFAULTS: Record<string, string> = {
   ...colorDefaults,
@@ -59,7 +77,7 @@ const ALL_DEFAULTS: Record<string, string> = {
   ...easeDefaults,
 };
 
-interface PlaygroundThemeEditorProps {
+interface ThemeEditorProps {
   /** Light/dark mode, owned by the playground so editor + preview stay in sync. */
   mode: 'light' | 'dark';
   /** Optional theme to seed the editor's token + component state from. */
@@ -68,37 +86,35 @@ interface PlaygroundThemeEditorProps {
   onThemeChange: (theme: XDSDefinedTheme) => void;
 }
 
-export function PlaygroundThemeEditor({
+export function ThemeEditor({
   mode,
   initialTheme,
   onThemeChange,
-}: PlaygroundThemeEditorProps) {
-  const [tokens, setTokens] = React.useState<Record<string, string>>(() => ({
+}: ThemeEditorProps) {
+  const [tokens, setTokens] = useState<Record<string, string>>(() => ({
     ...ALL_DEFAULTS,
     ...initialTheme?.tokens,
   }));
   // Component-level overrides from the seeded theme. Kept separate from token
   // state so they form the base layer that token + custom overrides compose on.
-  const [baseComponents] = React.useState<Record<string, unknown>>(
+  const [baseComponents] = useState<Record<string, unknown>>(
     () => initialTheme?.components ?? {},
   );
-  const [panelTab, setPanelTab] = React.useState<
-    'theme' | 'components' | 'tokens'
-  >('theme');
-  const [typeScaleBase, setTypeScaleBase] = React.useState(14);
-  const [typeScaleRatio, setTypeScaleRatio] = React.useState(1.2);
-  const [radiusBase, setRadiusBase] = React.useState(4);
-  const [spacingBase, setSpacingBase] = React.useState(4);
-  const [sizeBase, setSizeBase] = React.useState(32);
-  const [durationStep, setDurationStep] = React.useState(1);
+  const [panelTab, setPanelTab] = useState<'theme' | 'components' | 'tokens'>(
+    'theme',
+  );
+  const [typeScaleBase, setTypeScaleBase] = useState(14);
+  const [typeScaleRatio, setTypeScaleRatio] = useState(1.2);
+  const [radiusBase, setRadiusBase] = useState(4);
+  const [spacingBase, setSpacingBase] = useState(4);
+  const [sizeBase, setSizeBase] = useState(32);
+  const [durationStep, setDurationStep] = useState(1);
   // A unified preset is only "active" once the user explicitly applies one.
-  const [activePreset, setActivePreset] = React.useState<string | null>(null);
-  const [autoPickColors, setAutoPickColors] = React.useState(false);
-  const [customOverrides, setCustomOverrides] = React.useState<
-    CustomOverride[]
-  >([]);
+  const [activePreset, setActivePreset] = useState<string | null>(null);
+  const [autoPickColors, setAutoPickColors] = useState(false);
+  const [customOverrides, setCustomOverrides] = useState<CustomOverride[]>([]);
 
-  const currentTheme = React.useMemo(() => {
+  const currentTheme = useMemo(() => {
     const coreTokens: Record<string, string> = {};
     const componentTokens: Record<string, string> = {};
     for (const [key, value] of Object.entries(tokens)) {
@@ -125,50 +141,28 @@ export function PlaygroundThemeEditor({
 
   // Report the composed theme upward so the playground can push it to the
   // preview iframe whenever tokens or overrides change.
-  React.useEffect(() => {
+  useEffect(() => {
     onThemeChange(currentTheme);
   }, [currentTheme, onThemeChange]);
 
-  const handleTokenChange = React.useCallback((name: string, value: string) => {
+  const handleTokenChange = useCallback((name: string, value: string) => {
     setTokens(prev => ({...prev, [name]: value}));
   }, []);
 
-  const applyTypeScale = React.useCallback((base: number, ratio: number) => {
+  const applyTypeScale = useCallback((base: number, ratio: number) => {
     setActivePreset(null);
     setTypeScaleBase(base);
     setTypeScaleRatio(ratio);
     setTokens(prev => ({...prev, ...expandTypeScale({base, ratio})}));
   }, []);
 
-  const applySpacingScale = React.useCallback((base: number) => {
+  const applySpacingScale = useCallback((base: number) => {
     setActivePreset(null);
     setSpacingBase(base);
-    const steps = [0, 0.5, 1, 1.5, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12];
-    const keys = [
-      '0',
-      '0-5',
-      '1',
-      '1-5',
-      '2',
-      '3',
-      '4',
-      '5',
-      '6',
-      '7',
-      '8',
-      '9',
-      '10',
-      '11',
-      '12',
-    ];
-    const patch: Record<string, string> = {};
-    steps.forEach((step, i) => {
-      patch[`--spacing-${keys[i]}`] = `${Math.round(base * step)}px`;
-    });
-    setTokens(prev => ({...prev, ...patch}));
+    setTokens(prev => ({...prev, ...buildSpacingScale(base)}));
   }, []);
 
-  const applySizeScale = React.useCallback((base: number) => {
+  const applySizeScale = useCallback((base: number) => {
     setActivePreset(null);
     setSizeBase(base);
     setTokens(prev => ({
@@ -179,7 +173,7 @@ export function PlaygroundThemeEditor({
     }));
   }, []);
 
-  const applyRadiusScale = React.useCallback((base: number) => {
+  const applyRadiusScale = useCallback((base: number) => {
     setActivePreset(null);
     setRadiusBase(base);
     setTokens(prev => ({
@@ -188,7 +182,7 @@ export function PlaygroundThemeEditor({
     }));
   }, []);
 
-  const applyDurationScale = React.useCallback((multiplier: number) => {
+  const applyDurationScale = useCallback((multiplier: number) => {
     setDurationStep(multiplier);
     const defaults: Record<string, number> = {
       '--duration-fast-min': 130,
@@ -208,7 +202,7 @@ export function PlaygroundThemeEditor({
     setTokens(prev => ({...prev, ...patch}));
   }, []);
 
-  const applyUnifiedPreset = React.useCallback((presetKey: string) => {
+  const applyUnifiedPreset = useCallback((presetKey: string) => {
     const p = UNIFIED_PRESETS[presetKey as keyof typeof UNIFIED_PRESETS];
     if (!p) {
       return;
@@ -219,43 +213,18 @@ export function PlaygroundThemeEditor({
     setSpacingBase(p.spacing);
     setRadiusBase(p.radius);
     setSizeBase(p.sizeMd);
-    setTokens(prev => {
-      const spacingSteps = [0, 0.5, 1, 1.5, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12];
-      const spacingKeys = [
-        '0',
-        '0-5',
-        '1',
-        '1-5',
-        '2',
-        '3',
-        '4',
-        '5',
-        '6',
-        '7',
-        '8',
-        '9',
-        '10',
-        '11',
-        '12',
-      ];
-      const spacingPatch: Record<string, string> = {};
-      spacingSteps.forEach((step, i) => {
-        spacingPatch[`--spacing-${spacingKeys[i]}`] =
-          `${Math.round(p.spacing * step)}px`;
-      });
-      return {
-        ...prev,
-        ...expandTypeScale({base: p.typeBase, ratio: p.typeRatio}),
-        ...spacingPatch,
-        ...expandRadiusScale({base: p.radius, multiplier: 1}),
-        '--size-element-sm': `${p.sizeMd - p.spacing}px`,
-        '--size-element-md': `${p.sizeMd}px`,
-        '--size-element-lg': `${p.sizeMd + p.spacing}px`,
-      };
-    });
+    setTokens(prev => ({
+      ...prev,
+      ...expandTypeScale({base: p.typeBase, ratio: p.typeRatio}),
+      ...buildSpacingScale(p.spacing),
+      ...expandRadiusScale({base: p.radius, multiplier: 1}),
+      '--size-element-sm': `${p.sizeMd - p.spacing}px`,
+      '--size-element-md': `${p.sizeMd}px`,
+      '--size-element-lg': `${p.sizeMd + p.spacing}px`,
+    }));
   }, []);
 
-  const handleExpandColorScale = React.useCallback((accentHex: string) => {
+  const handleExpandColorScale = useCallback((accentHex: string) => {
     const derived = expandColorScale({accent: accentHex});
     let hex = accentHex.replace('#', '');
     // Normalize 3-char short hex (e.g. "f00") to 6-char ("ff0000")
@@ -277,19 +246,10 @@ export function PlaygroundThemeEditor({
   }, []);
 
   return (
-    <div
-      style={{
-        flex: 1,
-        minHeight: 0,
-        display: 'flex',
-        flexDirection: 'column',
-        overflow: 'hidden',
-      }}>
-      {/* Self-load theme fonts (mirrors ThemeEditorView) so the editor's
-          controls render in the loaded theme's typeface even when used
-          standalone. */}
+    <XDSVStack xstyle={s.root}>
+      {/* Self-load theme fonts so the editor's controls render in the loaded
+          theme's typeface even when used standalone. */}
       <style>{`@import url("${GOOGLE_FONTS_URL}");`}</style>
-      {/* Tab strip */}
       <XDSTabList
         hasDivider
         value={panelTab}
@@ -299,10 +259,9 @@ export function PlaygroundThemeEditor({
         <XDSTab value="tokens" label="Advanced" />
       </XDSTabList>
 
-      {/* Scrollable editor content */}
-      <div style={{flex: 1, overflow: 'auto', padding: 16}}>
+      <XDSStackItem size="fill" xstyle={s.body}>
         {panelTab === 'theme' && (
-          <EditorSections
+          <BaseStylesPanel
             tokens={tokens}
             mode={mode}
             typeScaleBase={typeScaleBase}
@@ -322,7 +281,6 @@ export function PlaygroundThemeEditor({
             onApplyUnifiedPreset={applyUnifiedPreset}
             onSetAutoPickColors={setAutoPickColors}
             onExpandColorScale={handleExpandColorScale}
-            onSetTokens={setTokens}
           />
         )}
         {panelTab === 'components' && (
@@ -341,7 +299,7 @@ export function PlaygroundThemeEditor({
             onTokenChange={handleTokenChange}
           />
         )}
-      </div>
-    </div>
+      </XDSStackItem>
+    </XDSVStack>
   );
 }
